backend/requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
sqlmodel==0.0.22
pydantic==2.9.2
pydantic-settings==2.6.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
sqlalchemy==2.0.35
aiosqlite==0.20.0
python-multipart==0.0.20
python-dotenv==1.0.1